implemecion de control de denegacion de edicion a otro usuario

+++ app/Http/Controllers/Api/ProfileController.php
@@ -34,6 +34,12 @@ class ProfileController
      */
     public function update(Request $request, int $userId)
     {
+        $user = auth()->user();
+
+        if ($user->id_usuario != $userId) {
+            return response()->json(['message' => 'No autorizado'], 403);
+        }
+
         $data = $request->only([
             'correo',
             'nombre',
@@ -59,6 +65,12 @@ class ProfileController
 
     public function updateVisibility(Request $request, int $userId)
     {
+        $user = auth()->user();
+
+        if ($user->id_usuario != $userId) {
+            return response()->json(['message' => 'No autorizado'], 403);
+        }
+
         $data = $request->all();
 
         return response()->json(
@@ -70,15 +82,21 @@ class ProfileController
      * Controller para manejar la subida de imágenes de perfil y banner.
      */
 
-    public function uploadImage(Request $request, $usuarioId)
+    public function uploadImage(Request $request, int $userId)
     {
+        $user = auth()->user();
+
+        if ($user->id_usuario != $userId) {
+            return response()->json(['message' => 'No autorizado'], 403);
+        }
+
         $request->validate([
             'tipo' => 'required|in:profile,banner',
             'file' => 'required|file|image|mimes:jpg,jpeg,png,webp|max:5120'
         ]);
 
         $response = $this->service->addImageProfileBanner(
-            $usuarioId,
+            $userId,
             $request->file('file'),
             $request->tipo
         );
@@ -90,14 +108,20 @@ class ProfileController
      * Controller para manejar la eliminación de imágenes de perfil y banner.
      */
 
-    public function deleteImage(Request $request, $usuarioId)
+    public function deleteImage(Request $request, int $userId)
     {
+        $user = auth()->user();
+
+        if ($user->id_usuario != $userId) {
+            return response()->json(['message' => 'No autorizado'], 403);
+        }
+
         $request->validate([
             'tipo' => 'required|in:profile,banner'
         ]);
 
         $response = $this->service->deleteProfileBannerImage(
-            $usuarioId,
+            $userId,
             $request->tipo
         );
 
@@ -108,20 +132,21 @@ class ProfileController
      * Controller para manejar la actualización de imágenes de perfil y banner.
      */
 
-    public function updateImage(Request $request, $usuarioId)
+    public function updateImage(Request $request, int $userId)
     {
-    /*dd([
-        'all' => $request->all(),
-        'files' => $request->allFiles(),
-        'file' => $request->file('file'),
-    ]);*/
+        $user = auth()->user();
+
+        if ($user->id_usuario != $userId) {
+            return response()->json(['message' => 'No autorizado'], 403);
+        }
+
         $request->validate([
             'tipo' => 'required|in:profile,banner',
             'file' => 'required|file|image|mimes:jpg,jpeg,png,webp|max:5120'
         ]);
 
         $response = $this->service->updateProfileBannerImage(
-            $usuarioId,
+            $userId,
             $request->file('file'),
             $request->tipo
         );
